Moves functions to a separate file
