+++ templates/config-setup.rb
@@ -1,19 +1,6 @@
 #!/usr/bin/env ruby
 require 'yaml'
-
-# Helper class for formatting message text
-class String
-  def red;     "\e[31m#{self}\e[0m" end
-  def green;   "\e[32m#{self}\e[0m" end
-  def yellow;  "\e[33m#{self}\e[0m" end
-  def bold;    "\e[1m#{self}\e[22m" end
-end
-
-# Helper function for printing error messages
-def print_error(message)
-   puts "===> Dreambox config: #{message}".bold.red
-   abort "     See 'Getting Started': https://github.com/goodguyry/dreambox/wiki".bold.yellow
-end
+require_relative 'config-utilities.rb'
 
 module Config
   vm_config_file = $vm_config_file
@@ -139,31 +126,6 @@ module Config
 
   # Debug formatting
   if VM_CONFIG['debug'] then
-    puts ''
-    puts "===> Dreambox Debug:".bold.yellow
-    puts "     Config File:   #{vm_config_file_path}".bold
-    puts "     Box Name:      #{VM_CONFIG['name']}".bold
-    puts "     PHP Version:   #{VM_CONFIG['php']}".bold
-    puts "     SSL Enabled:   #{VM_CONFIG['ssl_enabled']}".bold
-    if 0 < "#{VM_CONFIG['hosts']}".length then
-      puts "     Hosts:         #{VM_CONFIG['hosts']}".bold
-    end
-    puts ''
-    VM_CONFIG['sites'].each do |site, items|
-      puts "===> #{site}: username     #{items['username']}".bold
-      puts "     #{site}: root_path    #{items['root_path']}".bold
-      puts "     #{site}: root         #{items['root']}".bold
-      puts "     #{site}: local_root   #{items['local_root']}".bold
-      if 0 < "#{items['ssl']}".length then
-        puts "     #{site}: ssl          #{items['ssl']}".bold
-      end
-      if 0 < "#{items['aliases']}".length then
-        puts "     #{site}: aliases      #{items['aliases']}".bold
-      end
-      puts "     #{site}: box_name     #{items['box_name']}".bold
-      puts "     #{site}: host         #{items['host']}".bold
-      puts ''
-    end
-    puts ''
+    print_debug_info(VM_CONFIG, vm_config_file_path)
   end
 end
